refactor(core/directives): remove old code and start preparing bindings for 2.0
